auth/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
authlib==1.6.3
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.2.1
    # via uvicorn
cryptography==45.0.7
    # via authlib
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
structlog==25.4.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
